[FIX] Files hierarchy refactored
