+++ Makefile
@@ -33,7 +33,9 @@ UTILS_DIR	=	utils
 UTILS_SRC	=	lists.c \
 				logical.c \
 				redirections.c \
-				strings.c
+				strings.c \
+				history/history.c \
+				history/interact.c
 
 ENV_DIR	=	environment
 ENV_SRC	=	environment.c \
